several fix for linux & the replay

+++ release/distrib/replay/lib/include/synth.h
@@ -70,7 +70,8 @@ typedef unsigned __int64 Uint64;
 #else
 typedef long long int64;
 #if defined(__LINUX__) && !defined(__FREEBSD__)
-typedef unsigned long long uint64_t;
+typedef long unsigned int uint64_t;
+#include <sys/types.h>
 typedef uint64_t Uint64;
 #else
 typedef unsigned long long Uint64;
